Run with updated Flask API and Python 3 only
This change uses the post 1.0 version of the Flask API.  We drop
support for Python 2 which allows us to remove dependencies on
`mock` and `six`.  Furthermore, we test against the documented
`postevent` from GitHub.  We choose to use the entire object
rather than a `postevent`-like object in order to test against
"real world" data.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>